+++ routes/web.php
@@ -20,8 +20,8 @@ use App\Http\Controllers\PostCommentController;
 use App\Http\Controllers\CommentReactionController;
 use App\Http\Controllers\PostController;
 use App\Http\Controllers\PostReactionController;
-use App\Http\Controllers\PostReportController;
 use App\Http\Controllers\PrivacyController;
+use App\Http\Controllers\ReportController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\Profile\PublicProfileController;
 use App\Http\Controllers\SavedPostController;
@@ -96,7 +96,9 @@ Route::middleware(['auth', 'banned', 'track_last_seen'])->group(function () {
     Route::post('/posts/{post}/save', [SavedPostController::class, 'toggle'])->name('posts.save.toggle');
     Route::post('/posts/{post}/pin', [PostController::class, 'pin'])->name('posts.pin');
     Route::delete('/posts/{post}/pin', [PostController::class, 'unpin'])->name('posts.unpin');
-    Route::post('/posts/{post}/report', [PostReportController::class, 'store'])->name('posts.report');
+    Route::post('/posts/{post}/report', [ReportController::class, 'reportPost'])->name('posts.report');
+    Route::post('/posts/{post}/comments/{comment}/report', [ReportController::class, 'reportComment'])->name('comments.report');
+    Route::post('/users/{user:username}/report', [ReportController::class, 'reportUser'])->name('users.report');
 
     Route::get('/pets/create', [PetController::class, 'create'])->name('pets.create');
     Route::post('/pets', [PetController::class, 'store'])->name('pets.store');